fix: ignore spdx header for E2E tests
We want to add the SPDX headers for the test files, but we don't
generate them with our tools. Therefore, we special case these first
three lines in our E2E tests.

We do not want to do an 'in' in the tests, as it would also succeed if
we accidentally generate empty files.

Signed-off-by: Stefan Kraus <stefan.kraus@ul.com>