fix: handle upon 'or' case more properly
When given 3 or more 'or' events in the 'upon' clause,
the regular_condition_unexp_pt splits (a|b|c) into ((a|b)|c), surrounded
by brackets. This was not handled properly, the result contained e.g. (a
with bracket.

This shows that the current split behaviour is not sufficient for cases
where all of FRET allowed features is correctly transformed. While
sufficient for simpler cases and the features we used, there is still
work to do for a proper transformation case.

Signed-off-by: Stefan Kraus <stefan.kraus@ul.com>